docker-compose.yaml

+++ webook/internal/repository/dao/user.go
@@ -30,7 +30,7 @@ func (dao *UserDao) FindByEmail(ctx context.Context, email string) (User, error)
 	return u, err
 }
 
-func (dao *UserDao) FindById(ctx context.Context, id int) (User, error) {
+func (dao *UserDao) FindById(ctx context.Context, id int64) (User, error) {
 	var u User
 	err := dao.db.WithContext(ctx).Where("id=?", id).First(&u).Error
 	return u, err
